+++ graph/graph-adapter-output-spring-data-neo4j-sdn6/src/containerTest/kotlin/eu/tib/orkg/prototype/statements/adapter/output/neo4j/spring/SpringDataNeo4jListAdapterContractTests.kt
@@ -47,4 +47,10 @@ internal class SpringDataNeo4jListAdapterContractTests(
             springDataNeo4jStatementAdapter
         )
     )
+
+    finalizeSpec {
+        springDataNeo4jResourceAdapter.deleteAll()
+        springDataNeo4jStatementAdapter.deleteAll()
+        springDataNeo4jPredicateAdapter.deleteAll()
+    }
 })